python/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.3
openpyxl==3.1.2
schedule==1.2.1
python-dotenv==1.0.0
lxml==4.9.3
